any_storage

Module noop

Module for noop storage implementation.

Structs

NoopStore

Noop store, does nothing, can’t do anything

NoopStoreConfig

Configuration for NoopStore.

NoopStoreDirectory

Representation of a directory in the noop store.

NoopStoreDirectoryReader

Reader for streaming entries from a noop store directory.

NoopStoreFile

Representation of a file in the noop store.

NoopStoreFileMetadata

Metadata associated with a file in the store store.

NoopStoreFileReader

Reader for asynchronously reading the contents of a file in the noop store.

NoopStoreFileWriter

Writer for files in the noop store.

Type Aliases

NoopStoreEntry